backend/services/orchestrator/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0

# ElevenLabs SDK for carrier_outreach agent
elevenlabs>=1.5.0

# LangGraph and OpenTelemetry
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-httpx>=0.45b0
opentelemetry-exporter-otlp>=1.24.0


